refactor: 모임 개설 페이지의 컨텐츠 부분 컴포넌트화
