+++ src/app/[lng]/(main)/grouping/[groupId]/components/articles/ArticleSection.client.tsx
@@ -1,18 +1,18 @@
 'use client';
 
-import { useGetArticles, useGetGroupDetail } from '@/apis/groups/queries';
+import { GroupDetailResponse } from '@/apis/groups';
+import { useGetArticles } from '@/apis/groups/queries';
 import ArticleItem from '@/app/[lng]/(main)/grouping/components/ArticleItem.client';
 import { ItemList } from '@/components/List';
 import { useNumberParams } from '@/hooks/useNumberParams';
 import { useBlockStore } from '@/store/useBlockStore';
 
-export default function ArticleSection() {
+interface ArticlesContentProps extends GroupDetailResponse {}
+
+export default function ArticleSection({ isCaptain }: ArticlesContentProps) {
   const { blockArticleIds } = useBlockStore();
   const { groupId } = useNumberParams<['groupId']>();
   const { data: articlesData } = useGetArticles(groupId);
-  const { data: groupDetailData } = useGetGroupDetail(groupId);
-
-  const { isCaptain } = groupDetailData;
 
   return (
     <section>
