--- src/app/[lng]/(main)/grouping/[groupId]/components/articles/index.ts
@@ -1,2 +1,0 @@
-export { default as ArticleSection } from './ArticleSection.client';
-export { default as NoticeSection } from './NoticeSection.client';
